+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 - マニュアル等の不備を修正．
 
 ## 1.0.9 2019/05/30
-- 丸めを無視した多項式近似(変数は製作レベル)を表す `Mgmg::TPolynomial`およびこれを生成するメソッド String#poly を追加．
+- 丸めを無視した多項式近似(変数は製作レベル)を表す`Mgmg::TPolynomial`およびこれを生成するメソッド`String#poly`を追加．
 
 ## 1.0.10 2019/06/29
 - 消費エレメント量を返す`Mgmg::Equip#total_cost`を追加．
@@ -96,4 +96,4 @@
 
 ## 1.3.2
 - `Mgmg::TPolynomial`に比較演算子を追加．
-- `String#phydef_optimize`を追加．
+- `String#phydef_optimize`，`String#buster_optimize`を追加．